fix a few warnings on 64-bit machines
size_t is bigger than int on most 64-bit machines, so cast
size_t to long when passing them to printf-like functions.

Ideally we'd use %z, but many compilers don't support it.

git-svn-id: https://svn.musicpd.org/mpd/trunk@4656 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/interface.c
@@ -332,10 +332,10 @@ static int processLineOfInput(Interface * interface)
 				ERROR("interface %i: command "
 				      "list size (%i) is "
 				      "larger than the max "
-				      "(%i)\n",
+				      "(%li)\n",
 				      interface->num,
 				      interface->cmd_list_size,
-				      interface_max_command_list_size);
+				      (long)interface_max_command_list_size);
 				closeInterface(interface);
 				ret = COMMAND_RETURN_CLOSE;
 			} else
